# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cutting-room/rules": {
      "get": {
        "operationId": "CuttingRoomController_getRules",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      },
      "put": {
        "operationId": "CuttingRoomController_putRules",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCuttingRoomRulesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/libraries": {
      "get": {
        "operationId": "CuttingRoomController_listLibraries",
        "parameters": [
          {
            "name": "mediaType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/diskspace": {
      "get": {
        "operationId": "CuttingRoomController_diskspace",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/analyze": {
      "post": {
        "operationId": "CuttingRoomController_analyze",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAnalyzeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/snapshots": {
      "get": {
        "operationId": "CuttingRoomController_listSnapshots",
        "parameters": [
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/snapshots/{id}": {
      "get": {
        "operationId": "CuttingRoomController_getSnapshot",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/snapshots/{id}/candidates": {
      "get": {
        "operationId": "CuttingRoomController_listCandidates",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dir",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxTier",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minScore",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rootFolder",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "watchStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "selectedOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      },
      "patch": {
        "operationId": "CuttingRoomController_patchSelection",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSelectionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/snapshots/{id}/rootfolders": {
      "get": {
        "operationId": "CuttingRoomController_listSnapshotRootFolders",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/snapshots/{id}/selection/auto": {
      "post": {
        "operationId": "CuttingRoomController_autoSelect",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutoSelectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/snapshots/{id}/prune": {
      "post": {
        "operationId": "CuttingRoomController_prune",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartPruneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/snapshots/{id}/stop": {
      "post": {
        "operationId": "CuttingRoomController_stop",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/prunes": {
      "get": {
        "operationId": "CuttingRoomController_listPrunes",
        "parameters": [
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "restored",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/prunes/{id}/restore": {
      "post": {
        "operationId": "CuttingRoomController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/duplicates": {
      "get": {
        "operationId": "CuttingRoomController_listDuplicates",
        "parameters": [
          {
            "name": "sectionKey",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/duplicates/cleanup": {
      "post": {
        "operationId": "CuttingRoomController_duplicateCleanup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateCleanupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/large-files": {
      "get": {
        "operationId": "CuttingRoomController_listLargeFiles",
        "parameters": [
          {
            "name": "threshold",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mediaType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sectionKeys",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceIds",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/large-files/replace": {
      "post": {
        "operationId": "CuttingRoomController_largeFilesReplace",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LargeFilesReplaceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/wanted": {
      "get": {
        "operationId": "CuttingRoomController_listWanted",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
    "/api/cutting-room/wanted/prune": {
      "post": {
        "operationId": "CuttingRoomController_wantedPrune",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WantedPruneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "cutting-room"
        ]
      }
    },
...
  },
  "info": {
    "title": "Immaculaterr API",
    "description": "Local API for the Immaculaterr webapp.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpdateCuttingRoomRulesDto": {
        "type": "object",
        "properties": {}
      },
      "StartAnalyzeDto": {
        "type": "object",
        "properties": {}
      },
      "AutoSelectDto": {
        "type": "object",
        "properties": {}
      },
      "PatchSelectionDto": {
        "type": "object",
        "properties": {}
      },
      "StartPruneDto": {
        "type": "object",
        "properties": {}
      },
      "DuplicateCleanupDto": {
        "type": "object",
        "properties": {}
      },
      "LargeFilesReplaceDto": {
        "type": "object",
        "properties": {}
      },
      "WantedPruneDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}